updating images adding some filters
