+++ package.json
@@ -15,7 +15,7 @@
     "babel-loader": "8.0.5",
     "babel-plugin-named-asset-import": "^0.3.4",
     "babel-preset-react-app": "^7.0.1",
-    "bento-components": "^0.1.20",
+    "bento-components": "^0.1.24",
     "bfj": "6.1.1",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
     "classnames": "^2.2.6",
