+++ public/index.html
@@ -7,7 +7,7 @@
     <meta name="apple-mobile-web-app-capable" content="yes" />
     <meta charset="utf-8" />
     <meta name="robots" content="noindex, nofollow" />
-    <link rel="shortcut icon" href="https://raw.githubusercontent.com/CBIIT/ctdc-codebase/master/src/main/frontend/public/favicon.ico" />
+    <link rel="shortcut icon" href="https://raw.githubusercontent.com/CBIIT/bento-frontend/master/public/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <!--
