+++ src/bento/headerData.js
@@ -3,7 +3,7 @@
 // globalHeaderLogo image 468x80
 // globalHeaderImage: image 1200x60
 export default {
-  globalHeaderLogo: 'https://raw.githubusercontent.com/CBIIT/bento-frontend/master/src/assets/header/icdc_nih_logo.svg',
+  globalHeaderLogo: 'https://raw.githubusercontent.com/CBIIT/bento-tools/master/src/components/assets/header/CTDC_Logo.svg',
   globalHeaderLogoLink: 'https://frederick.cancer.gov/',
   globalHeaderLogoAltText: 'Bento Logo',
   // globalHeaderImage: '',
